Remove WAMPv2 things, let's use plain websockets.
The RPC & pubsub bits of WAMP are awesome, but its P2P architecture is a poor
fit for what is essentially a client/server app. Requiring a WAMP router
involves a lot of extra moving parts/plumbing with twisted and complicates
authentication, which seems like a poorly defined area of WAMPv2 anyways (I
thought wamp-cra looked nice, but that's v1 only).